working emulator prototpye at 60fps

+++ samples/nes_emulator/src/virtual_screen.cpp
@@ -1,7 +1,7 @@
 #include "virtual_screen.h"
 #include <string.h>
 
-uint32_t pixels[256 * 240];
+uint8_t pixels[256 * 240];
 volatile bool RENDER_ENABLED;
 
 void screen_init(void)
